Rework the top bar
