--- src/ui/global-nav.module.css
@@ -1,63 +1,0 @@
-.globalNavContainer {
-    width: 100%;
-    background-color: lightslategray;
-    color: white;
-    padding: 1em;
-}
-
-.menuList {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-
-.navItemContainer {
-    list-style-type: none;
-    margin-left: 1em;
-    font-weight: 200;
-}
-
-.navItemContainer.active {
-    font-weight: 400;
-}
-
-.navItemGroup {
-    display: flex;
-    flex-direction: row;
-}
-
-.label {
-    background-color: transparent;
-    border: none;
-    font-weight: 200;
-    color: white;
-}
-
-.menuButton {
-    background-color: transparent;
-    border: none;
-    font-weight: 200;
-    color: white;
-    padding: 0;
-    cursor: pointer;
-}
-
-.popup {
-    position: absolute;
-    background-color: #9a6cb4;
-    color: white;
-    padding: 1em;
-    margin-top: 0.5em;
-    margin-left: -0.5em;
-    border-radius: 0.4em;
-    font-weight: 200;
-}
-
-.menuLink {
-    font-weight: 200;
-}
-
-.balance {
-    display: inline;
-    color: yellow;
-}
